10/7/2023 Update 0.2
Polished the UI and various entity sprite assets, including icons for tabs and
Added a small selection of sounds
Increased the chance of dropping scrap significantly for all enemies
Increased the player inventory by one slot
Decreased the spread and damage of the shotgun
Added a maximum range to the player, and an arrow that points to the mothership
Fixed bug #10
Fixed bug #11
Swapped the smelting tab and the garage tab
Expanded the size of the garage menu significantly
Added a wave counter to the player menu.
Changed the name of the game from Space Wars to Lagrange Commander
Added a proper title screen, and expanded the victory screen to include your time.
Added the training simulator mode